Added deriving Data, Typeable to some classes.
Need to information in advance of compiling to SQL on which tables
are used. Collect this information from combinators using SYB (yes this
is SYB at runtime, but for now it's `the easy' fix, if it works we might
want to pick something more efficient!).

The table information is needed for the column names, generating
arbitrary column names is not appreciated by the DB.