Changing detupler.
Fst and Snd are no longer eliminated in detupling phase but in the vec2algebra phase. This allows us to change the way we represent tuples without having to restructure the compiler.

Removed detupling behaviour for length. Instead the length primitive can now deal with tuples (that is list of tuples)