Modification to the CL rewriting strategy: We always prefer flat joins instead of nestjoins. Nestjoins are rather a maybe-not-so-bad exist strategy. As organisation-nested.Q2 demonstrates, it is sometimes necessary to normalize childs of the current node (esp. guards) before flat joins can be introduced. However, if we interleave nestjoin introduction with flatjoins/normalization, nestjoins might be introduced too early and block introduction of e.g. semijoins and antijoins.
Here, we try to separate flat and nested joins into two separate phases. First,
we perform a top-down-traversal during which cleanups, comprehension
normalization rules and flat join rules are applied. The goal here is to
exhaustively introduce every flat join we can. In a separate phase, we perform a
bottom-up traversal and only look for nestjoins. Of course, this might lead to
further opportunities for flat joins and unnesting being uncovered. Therefore,
we repeat both phases when changes were made.

This is rough. It is propably necessary to interleave the nestjoin traversal
with cleanups.