Re-implemented vectorization of singleton alignment.
With natural keys, we can not rely on singleton vectors in SShapes to
have the same key (1). Key values might be different and keys might not
be compatible at all. Instead of making keys uniform, we use CartProduct
for combining singleton vectors. CartProduct produces a composite key in
the result vector. On the positive side, we don't need to employ a join
predicate for the aligning join which was useless anyway. On the
negative side, because of the composite key, inner vectors must be
re-aligned.