UnboxSng produces a rekeying vector to align inner vectors.
Originally, UnboxSng (formerly UnboxScalar) was used only to unbox
aggregation results (AggrS). In this case, we could be sure that the
vector to be unboxed would not be referenced by any inner vectors
because aggregation results are scalar. However, UnboxSng is now also
used to unbox singletons for OnlyL. In this case, we might have inner
vectors which need to be brought to the key domain of the outer vector.