GroupAggr operates on individual segments.
The vector operator GroupAggr combines grouping and aggregation. In the
past, GroupAggr was regarded as a combination of Group and Aggrs. That
is, a vector with no/one segment was grouped globally and the individual
groups were aggregated. We change the semantics of GroupAggr to
correspond to a combination of GroupS and AggrS: In general, we expect a
segmented vector as input (it may consist only of the unit segment).
Each segment is grouped and aggregated individually. The result vector
has the same segment structure as the input vector. Each segment
contains a list of aggregate results for the individual groups computed
from the segment.

Rewrites that deal with GroupAggr (introduction, merging) are changed to
consider GroupS instead of Group.