Eliminate non-segment ops Number, Reverse, Unique.
These operators can be replaced by their per-segment counterparts. The
only overhead we suffer is sorting and grouping on constant segment
identifiers which can be easily removed.

On the level of VL rewrites, we incur one regression: Window rewrites
were based on the flat Number operator and are currently disabled. It
should be possible to transplant them to the per-segment NumberS
operator.