Keep vector representation uniform when unnesting.
Vectors are always divided into segments, even if we only have the unit
segment. Vectorizing operators that only pull inner vectors to the
outside (concat, only) we have to put those vectors into the unit
segment to make sure that the vector representation stays uniform. For
this reason we reintroduce the Unsegment operator that merges all
segments into the unit segment.