Eliminate operators NestJoin/Product, CartProduct.
These per-vector operators are replaced by their per-segment
counterparts.

For some rewrites it is not clear whether they can be generalized to
segmented input vectors. To avoid rewriting regressions, we introduce a
new property to determine if an input vector is actually flat (i.e. has
only the unit segment). In those cases, we actually have the old flat
case where the rewrites apply as before.