Introduce a small GroupJoin.
The switch to segment-only operators caused one regression: NestJoin
chains could no longer be rewritten because NestJoin now considers the
segment structure. To avoid this, we introduced an Unsegment in the
input of the second NestJoin to temporarily flatten the vector. Although
this works, we end up with another problem: The Unsegment blocks
introduction of a GroupJoin operator for the second NestJoin. This is
hard to avoid because the segment structure simply does not fit anymore.

The inability to introduce a GroupJoin on the top of a NestJoin chain
seriously impacts query performance in nested benchmarks (e.g.
custRevenues, expectedRevenueFor, topOrdersPerCust).

This is an attempt to attenuate the performance regression. We introduce
a "small" GroupJoin that does not perform unboxing. Without unboxing, we
do not need to check whether the left NestJoin input is identical with
the left Unbox input or whether they have the same segment structure.
The Unbox operator will stay as it is, as will the potential Unsegment
operator. AggrS and NestJoin are merged into one. This avoids at least
the default value computation in AggrS and the implied plan sharing.