Filter GroupJoin inputs in a NestJoin cascade.
If the right child of a NestJoin is a GroupJoin, we can not reorder the
corresponding vector operators. Consequently, the GroupJoin will be
executed for all elements from its left input, even if those will not
match any elements in the outer left input.

We attempt to ease this problem by applying a form of side-ways
information passing: Before performing the GroupJoin, its left input is
filtered to retain only those elements which will find a match in the
outermost vector.