Re-use aggregates when merging GroupJoins.
When an aggregate that is to be merged into an existing GroupJoin
already exists on the same GroupJoin, reuse the existing one. Note that
this will never trigger currently because the comparison of the
aggregates is too precise: The input types do not match. We need to only
compare the expressions without type annotations.