Remove incorrect vector rewrite.
On vector plans, we can not eliminate CartProductS operators for which
one side is constant and has cardinality 1. Although the same rewrite
would be valid on relational algebra, CartProductS changes the key
domain. Eliminating the operator and replacing it with a projection
removes this change of the key domain. This leads to inner and outer
vectors not being aligned.