Fix introduction of nestjoins in guards.
Nestjoin introduction used anytdR to search for a viable combination of
generator and guard for a nestjoin. This also matches generators of a
nested comprehension. Actually, we only want to consider generators of
the current comprehension. The fix replaces anytdR with a walk of the
outer comprehensions' qualifier list spine.