Eliminate nestproduct operator.
Nestproduct has been used for two use cases: as a fallback during
decorrelation when we can not use a nestjoin operator (due to lack of a
suitable join predicate). In this case, a nestproduct has the benefit of
avoiding cross-level references (and thus `dist`) by providing outer
values in the inner vector. However, this is a weak motivation. If we
are forced to introduce a nestproduct, decorrelation has essentially
failed and we are dead.

Additionally, the vector nestproduct operator has been used to replicate
vectors (i.e. implement dist^0 on a vector). Our plans for a new
decorrelation scheme heavily rely on the ability to not execute a dist
but have it produce virtual segments in a delayed vector representation.
Using nestproduct, this is not possible: Because nestproduct's defintion
requires to replicate values of the left input in the inner vector, we
can not delay the operator but are forced to materialize.

With this patch, we do not use nestproduct during decorrelation anymore.
To implement replication of vectors, we use a new operator
replicatevector that does not produce values from its left input and can
be delayed.