Generalize interface for algebra translation.
Previously, translation from a vector language to an algebraic plan
returned the vector language-specific monad. The relational backend
would use this computation to insert Serialize operators. With this API
however, it is not possible to re-use the Serialize code for multiple
vector languages (i.e. relational implementation of SL and VSL).

Here, we return the computation in the generic Build monad, stripping
away the layer specific to vector languages.