Replace rekeying map with replication map in zip.
We want to establish the property that application of a rekeying
map (KVec) does not remove entries from a vector and simply maps the
key. The use of rekeying maps in the zip implementation is the only one
that does not conform to this property: zip may remove dangling tuples
from the left or right.

Instead, zip now uses a replication map.