Add separate operator for segment merging.
We add a special operator for the implementation of lifted concat. Here, the
innermost segments have to be merged according to the segment structure of the
middle vector.

Formerly, this was implemented based on the rekeying mechanism. However, in the
unordered backend, order maintenance is not possible quite correctly with
rekeying: we need to combine the order labels of the middle vector with those
of the inner vector to guarantee the correct element order in the merged
segments.

As an additional benefit, this simplifies rekeying maps considerably.